+++ pong/Wall.h
@@ -23,7 +23,7 @@ struct WallCollider : RectangularColliderComponent
         m_manager = manager_;
     }
 
-    void update(float deltatime) override
+    void update(uint32_t deltatime) override
     {
         using fmt::format;
         auto transformComponent = m_owner->component<TransformComponent>();
@@ -46,7 +46,7 @@ struct WallDrawable : DrawableComponent
         m_transformComponent = owner->component<TransformComponent>();
     }
 
-    void update(float deltatime) override
+    void update(uint32_t deltatime) override
     {
         m_shape.setWidth(m_transformComponent->size().width());
         m_shape.setHeight(m_transformComponent->size().height());
